requirements.txt
streamlit>=1.28.0
snowflake-connector-python>=3.0.0
pandas>=2.0.0
plotly>=5.17.0
python-dotenv>=1.0.0
